Backend/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
ansicon==1.89.0
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==3.2.0
blessed==1.20.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.0.0
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contexttimer==0.3.3
cryptography==41.0.5
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.1.1
django-cors-headers==4.2.0
django-cron==0.6.0
django-inertia==1.3.0
django-js-routes==0.2.0
django-restframework==0.0.1
django-vite==2.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dotty-dict==1.3.0
duplicity==0.8.21
fasteners==0.14.1
filelock==3.12.3
frozenlist==1.4.0
future==0.18.2
gitdb==4.0.11
GitPython==3.1.40
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==1.1.2
gyp==0.1
httplib2==0.20.2
humanfriendly==10.0
idna==3.4
importlib-metadata==4.6.4
inertia-django==0.5.3
jeepney==0.7.1
jinxed==1.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
marshmallow==3.20.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
numpy==1.25.2
nvm==1.0.21
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
packaging==23.1
pandas==2.1.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipenv==2023.9.1
platformdirs==3.10.0
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyDrive==1.3.1
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyreadline3==3.4.1
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
pytz==2023.3.post1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
rsa==4.8
SecretStorage==3.3.1
setuptools-scm==8.0.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==1.4.31
sqlparse==0.4.4
srsly==2.4.8
systemd-python==234
tomli==2.0.1
twilio==8.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==2.0.5
usb-creator==0.3.7
virtualenv==20.24.4
wadllib==1.3.6
wcwidth==0.2.8
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
